Address Copilot code review feedback
- Fix findLatestByPlayerIds queries to use MAX(gameDate) instead of
  MAX(id) to correctly handle backfilled games
- Use Math.toIntExact() for long→int cast to fail fast on overflow
- Use fixed date in mock handler for deterministic test behavior

+++ backend/src/main/java/com/mlbstats/api/service/FavoriteApiService.java
@@ -137,8 +137,8 @@ public class FavoriteApiService {
                 playerDashboards,
                 totalTeamCount > MAX_DASHBOARD_TEAMS,
                 totalPlayerCount > MAX_DASHBOARD_PLAYERS,
-                (int) totalTeamCount,
-                (int) totalPlayerCount
+                Math.toIntExact(totalTeamCount),
+                Math.toIntExact(totalPlayerCount)
         );
     }
